refactor(client/pages): implement page component pattern for better separation of concerns
This commit introduces a cleaner architectural approach by moving page component
logic from the Next.js app directory into dedicated feature modules:

WHAT CHANGED:
- Created dedicated page components in feature directories:
  - Added `features/note/pages/` with note.page.tsx, create-note.page.tsx, and update-note.page.tsx
  - Added `features/video/pages/` with video.page.tsx
- Refactored app route files to be thin wrappers that import and use these components:
  - Updated notes/[noteId]/page.tsx
  - Updated notes/create/[videoId]/page.tsx
  - Updated notes/update/[noteId]/page.tsx
  - Updated videos/[videoId]/page.tsx

WHY:
- Improves separation of concerns by keeping business logic in features directory
- Enhances maintainability by grouping related functionality together
- Makes testing easier with more isolated components
- Provides clearer boundaries between routing and actual component implementation
- Follows the pattern already established for other parts of the application

This approach makes the codebase more scalable as new features and pages are added,
while keeping the Next.js app directory focused purely on routing.
