+++ apps/client/src/app/(dashboards)/notes/[noteId]/page.tsx
@@ -1,80 +1,13 @@
-"use client";
+import { NotePage } from "@/features/note/pages";
 
-import { useGetNoteById } from "@/features/note/hooks";
-import { useToggleVideoPlayer } from "@/hooks";
-
-import { Loader, MarkdownViewer, ResizablePanels } from "@/components/global";
-
-import {
-  NoteError,
-  NotePageFooter,
-  NotePageHeader,
-} from "@/features/note/components";
-
-import { useNoteStore } from "@/features/note/store";
-import { VideoPlayer } from "@/features/video/components";
-import { useEffect } from "react";
-
-type NotePageParams = {
-  noteId: string;
-};
-
-function NotePage({ params }: { params: NotePageParams }) {
-  const { noteId } = params;
-  const { noteActions } = useNoteStore();
-  const { data: note, isLoading, isError, refetch } = useGetNoteById(noteId);
-  const { isVideoPlayerVisible, toggleVideoPlayer } = useToggleVideoPlayer();
-
-  useEffect(() => {
-    noteActions.setNote(note);
-  }, [note, noteActions]);
-
-  if (isLoading) {
-    return (
-      <div className="min-h-screen flex items-center justify-center container max-w-4xl mx-auto px-4 py-8">
-        <Loader />
-      </div>
-    );
-  }
-
-  if (isError) {
-    return (
-      <div className="min-h-screen flex items-center justify-center container max-w-4xl mx-auto px-4 py-8">
-        <NoteError onRetry={() => refetch()} />
-      </div>
-    );
-  }
-
-  if (!note) return null;
-
-  return (
-    <main className="min-h-screen bg-white">
-      {/* Header */}
-      <NotePageHeader
-        noteId={note.id}
-        noteTitle={note.title}
-        isVideoVisible={isVideoPlayerVisible}
-        onToggleVideo={toggleVideoPlayer}
-      />
-
-      {/* Content */}
-      <article className="container h-screen mx-auto px-2 py-6 overflow-auto">
-        {isVideoPlayerVisible ? (
-          <ResizablePanels
-            leftSideContent={
-              <MarkdownViewer content={note.content} noteTitle={note.title} />
-            }
-            rightSideContent={<VideoPlayer videoId={note.youtubeId} />}
-          />
-        ) : (
-          <MarkdownViewer content={note.content} noteTitle={note.title} />
-        )}
-      </article>
-
-      {/* Footer */}
-      <NotePageFooter />
-    </main>
-  );
+interface IPageProps {
+  params: Promise<{ noteId: string }>;
 }
 
-export default NotePage;
+async function Page({ params }: IPageProps) {
+  const { noteId } = await params;
+
+  return <NotePage noteId={noteId} />;
+}
+
+export default Page;
