+++ apps/client/src/features/note/pages/create-note.page.tsx
@@ -1,0 +1,44 @@
+"use client";
+
+import { useCreateNote } from "@/features/note/hooks";
+import { useGetVideoById } from "@/features/video/hooks";
+
+import { NotePageLayout } from "@/features/note/components";
+import { useVideoStore } from "@/features/video/store";
+
+interface IPageProps {
+  videoId: string;
+}
+
+export function CreateNotePage({ videoId }: IPageProps) {
+  const { mutate: createNote, isPending: isSavingNote } = useCreateNote();
+  const { videoCurrentTime } = useVideoStore();
+  const { data: videoData, isLoading } = useGetVideoById(videoId);
+
+  const handleCreateNote = (noteTitle: string, content: string) => {
+    if (!videoData) return;
+
+    createNote({
+      videoId: videoData.id,
+      createNoteData: {
+        title: noteTitle,
+        content: content,
+        thumbnail: videoData.thumbnails.medium.url,
+        videoTitle: videoData.title,
+        youtubeId: videoData.youtubeId,
+        timestamp: videoCurrentTime,
+      },
+    });
+  };
+
+  return (
+    <NotePageLayout
+      videoId={videoId}
+      isLoading={isLoading || !videoData}
+      isSavingNote={isSavingNote}
+      modalTitle="Confirm Save Note"
+      modalDescription="Are you sure you want to save this note?"
+      handleSaveNote={handleCreateNote}
+    />
+  );
+}
