+++ apps/client/src/features/note/pages/index.ts
@@ -1,0 +1,3 @@
+export * from "./note.page";
+export * from "./update-note.page";
+export * from "./create-note.page";
