+++ apps/client/src/features/video/pages/index.ts
@@ -1,0 +1,1 @@
+export * from "./video.page";
